pip-requirements.txt
MouseInfo
PyGetWindow
PyMsgBox
PyYAML == 6.0.1
configparser
et-xmlfile
namegenerator
names
numpy
opencv-python
openpyxl
requests
robotframework
robotframework-excellib
robotframework-imaplibrary2
robotframework-requests
robotframework-seleniumlibrary
robotframework-archivelibrary
robotframework-datadriver
selenium
setuptools
urllib3
webdriver-manager
chromedriver
pandas
msedge-selenium-tools
pyaml-env
rpaframework
rpaframework-core
jsonpath
pycryptodome